(WIP) removed odd interface, reformat code
