+++ src/main/java/sem_06_notebook/models/Fileable.java
@@ -2,5 +2,6 @@ package sem_06_notebook.models;
 
 public interface Fileable<T> {
     T read();
+
     void save(T t);
 }
